requirements.txt
requests
# PyYAML
aiohttp
ruamel.yaml